+++ run_main.py
@@ -1,5 +1,4 @@
 from gitonic.main import main_func
 
 if __name__ == "__main__":
-
     main_func(debug_=True)