+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - added F5, F6 hotkeys for changing the active tabs
 - custom formatter support
   - defined under settings `formatter.json`
-    - if no file is found it defaults to `{".py": {"cmd": "black", "para": "%f"}}`
+-
 
 
 # version v0.0.10 - 20230225